mrcroqueto/requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.23
botocore==1.34.23
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
dj-static==0.0.6
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-import-export==3.3.5
django-js-asset==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg==3.1.16
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
static3==0.7.0
swagger-spec-validator==3.0.3
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
